・新MapData読み込み処理完了 ・合成種の触覚の更新完了

+++ CaseStudy/CaseStudy/res/h/Object/Flower.cpp
@@ -45,7 +45,7 @@ void CFlower::Init(D3DXVECTOR3 pos,D3DXVECTOR3 dir)
 
 	m_angle = AngleOf2Vector(pos,D3DXVECTOR3(0,1,0));
 
-	RotateZ(m_angle);
+	RotateZ((float)m_angle);
 
 	// アニメーション初期化
 	StartAnimation();
